Split the groups of tests into their own files.
