+++ src/test/jurisdictionVersions.spec.ts
@@ -2,26 +2,13 @@ import { expect } from "chai"
 import { describe, it, before } from "mocha"
 import { GetTestData, HitEndpoint } from '../dataCollection'
 import { LogArrayDifference } from '../dataCommunication'
-import { GetUniqueArrayOfIpcStates, ArrayDifference, GetUniqueArrayOfPaShimStates, GetUniqueArrayOfPaShimProducts } from '../dataManipulation'
-import { UsState, UsStateMapping, JurisdictionVersion, Jurisdiction, ProductLine, TestData } from '../interfaces/interfacesAndTypes'
+import { GetUniqueArrayOfIpcStates, ArrayDifference, GetUniqueArrayOfPaShimStates } from '../dataManipulation'
+import { UsState, UsStateMapping, JurisdictionVersion, TestData } from '../interfaces/interfacesAndTypes'
 import { Url, Endpoint } from '../constants'
 import shuffle from 'shuffle-array'
 
 let PA_SHIM_STATES_DATA:UsStateMapping;
 let JURISDICTION_VERSION_DATA:JurisdictionVersion[]
-let JURISDICTION_DATA:Jurisdiction[]
-let PRODUCT_LINE_DATA:ProductLine[]
-
-describe("~~~ HEALTH ~~~", () => {
-    it("Ensure that the /health endpoint works and is ok.", async () => {
-        const health:string = await HitEndpoint(Url.Dev, Endpoint.Health) as string
-
-        //testing
-        console.log(`health: ${health}`)
-
-        expect(health.toLowerCase()).to.equal("ok")
-    });
-})
 
 describe("~~~ JURISDICTION VERSION ~~~", () => {
     before("Get the test data.", async () => {
@@ -177,72 +164,4 @@ describe("~~~ JURISDICTION VERSION ~~~", () => {
     })
 })
 
-describe("~~~ JURISDICTION ~~~", () => {
-    before("Get the test data.", async () => {
-        const testData:TestData = await GetTestData(Url.Dev, Endpoint.Jurisdictions)
-        PA_SHIM_STATES_DATA = testData.paShimStatesData
-        JURISDICTION_DATA = testData.ipcData as Jurisdiction[]
-    });
-
-    describe("[CM-729] - Jurisdictions - 1 jurisdiction per state.", () => {
-        it("Verify that there is exactly 1 jurisdiction per state.", () => {
-            // Get list of unique ipc jurisdictions
-            const ipcJurisdictionUniqueNames:string[] = JURISDICTION_DATA.map((entry) => {
-                return entry.unique_name.replace("US-","")
-            })
-    
-            console.log("IPC jurisdiction Names (array)")
-            console.log(ipcJurisdictionUniqueNames)
-
-            console.log("IPC jurisdiction Names (set)")
-            console.log([...new Set(ipcJurisdictionUniqueNames)])
-    
-            // If the original arrary's length is equal to the number of its unique elements, then the original array only contained unique elements. In other words, no duplicates.
-            expect(ipcJurisdictionUniqueNames.length).to.equal([...new Set(ipcJurisdictionUniqueNames)].length)
-        })
-    
-        it("Verify that IPC has the same jurisdictions as pa_shim.", () => {
-            // Get list of unique ipc jurisdictions
-            const ipcJurisdictionUniqueNames:UsState[] = GetUniqueArrayOfIpcStates(JURISDICTION_DATA)
-    
-            // Get list of unique pa_shim jurisdictions
-            const paShimJurisdictionUniqueNames:UsState[] = GetUniqueArrayOfPaShimStates(PA_SHIM_STATES_DATA)
-    
-            // If the difference between the arrays is 0, then they contain an identical set of jurisdictions.
-            const jurisdictionArrayDifference:string[] = ArrayDifference(ipcJurisdictionUniqueNames, paShimJurisdictionUniqueNames)
-    
-            console.log("paShimJurisdictionUniqueNames")
-            console.log(paShimJurisdictionUniqueNames)
-            console.log("ipcJurisdictionUniqueNames")
-            console.log(ipcJurisdictionUniqueNames)
-    
-            expect(jurisdictionArrayDifference.length).to.equal(0)
-        })
-    })
-})
-
-describe("~~~ PRODUCT LINES ~~~", () => {
-    before("Get the test data.", async () => {
-        const testData:TestData = await GetTestData(Url.Dev, Endpoint.ProductLines)
-        PA_SHIM_STATES_DATA = testData.paShimStatesData
-        PRODUCT_LINE_DATA = testData.ipcData as ProductLine[]
-    });
-
-    describe("[CM-730] - Product Lines - IPC matches pa_shim.", () => {
-        it("Verify that the IPC products match the pa_shim products.", () => {
-            const ipcProductLineUniqueNames:string[] = PRODUCT_LINE_DATA.map((entry) => {return entry.unique_name})
-            const paShimProductsUniqueNames:string[] = GetUniqueArrayOfPaShimProducts(PA_SHIM_STATES_DATA)
-    
-            console.log("ipcProductLineUniqueNames")
-            console.log(ipcProductLineUniqueNames)
-            console.log("paShimProductsUniqueNames")
-            console.log(paShimProductsUniqueNames)
-    
-            // If the difference between the arrays is 0, then they contain an identical set of jurisdictions.
-            const productArrayDifference:string[] = ArrayDifference(ipcProductLineUniqueNames, paShimProductsUniqueNames)//ipcProductLineUniqueNames.filter((product) => !paShimProductsUniqueNames.includes(product));
-    
-            expect(productArrayDifference.length).to.equal(0)
-        })
-    })
-})
